+++ tests/e2e/game/game-options.test.js
@@ -73,6 +73,7 @@ describe("K - Game options", () => {
                         bearTamer: { doesGrowlIfInfected: false },
                         stutteringJudge: { voteRequestsCount: 2 },
                         wildChild: { isTransformationRevealed: true },
+                        dogWolf: { isChosenSideRevealed: true },
                         thief: { mustChooseBetweenWerewolves: false },
                         raven: { markPenalty: 3 },
                     },
@@ -94,6 +95,7 @@ describe("K - Game options", () => {
                 expect(game.options.roles.bearTamer.doesGrowlIfInfected).to.be.false;
                 expect(game.options.roles.stutteringJudge.voteRequestsCount).to.equal(2);
                 expect(game.options.roles.wildChild.isTransformationRevealed).to.be.true;
+                expect(game.options.roles.dogWolf.isChosenSideRevealed).to.be.true;
                 expect(game.options.roles.thief.mustChooseBetweenWerewolves).to.be.false;
                 expect(game.options.roles.raven.markPenalty).to.equal(3);
                 done();